@@ -13,7 +13,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <activemq.version>5.18.3</activemq.version>
+        <activemq.version>6.0.0</activemq.version>
         <junit.version>5.10.1</junit.version>
         <log4j.version>3.0.0-alpha1</log4j.version>
         <jackson-databind.version>2.16.0-rc1</jackson-databind.version>
@@ -206,7 +206,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <release>11</release>
+                    <release>17</release>
                 </configuration>
             </plugin>
             <plugin>
@@ -222,7 +222,6 @@
                 <configuration>
                     <parallel>all</parallel>
                     <useUnlimitedThreads>true</useUnlimitedThreads>
-                    <jvm>/Library/Java/JavaVirtualMachines/adoptopenjdk-11-openj9.jdk/Contents/Home/bin/java</jvm>
                 </configuration>
             </plugin>
             <plugin>
